requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pydantic
python-dotenv
gitpython
pydriller
